src/test/java/com/emse/spring/automacorp/api/RoomControllerTest.java
package com.emse.spring.automacorp.api;

import com.emse.spring.automacorp.dao.BuildingDao;
import com.emse.spring.automacorp.dao.HeaterDao;
import com.emse.spring.automacorp.dao.RoomDao;
import com.emse.spring.automacorp.dao.WindowDao;
import com.emse.spring.automacorp.model.BuildingEntity;
import com.emse.spring.automacorp.model.RoomEntity;
import com.emse.spring.automacorp.model.SensorEntity;
import com.emse.spring.automacorp.model.SensorType;
import com.fasterxml.jackson.databind.ObjectMapper;
import org.hamcrest.Matchers;
import org.junit.jupiter.api.Test;
import org.mockito.Mockito;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
import org.springframework.boot.test.mock.mockito.MockBean;
import org.springframework.http.MediaType;
import org.springframework.security.test.context.support.WithMockUser;
import org.springframework.test.web.servlet.MockMvc;
import org.springframework.test.web.servlet.request.MockMvcRequestBuilders;
import org.springframework.test.web.servlet.result.MockMvcResultMatchers;

import java.util.List;
import java.util.Optional;

import static org.mockito.BDDMockito.given;
import static org.springframework.http.MediaType.APPLICATION_JSON;
import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.*;
import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.*;


@WebMvcTest(RoomController.class)
public class RoomControllerTest {
    /**
     * All the controller tests were implemented before adding the authentication part of the backend.
     * The tests worked before the addition but the modifications made on a mock test don't work.
     * So in order to privilege security over the unit testing (knowing the real testing does work),
     *  I decided to keep the unit test as they were before (a mock test attempt can be found at the end of
     *  RoomControllerTest class).
     * To use the unit testing without the added security level, you have to remove/comment :
     *  - from build.gradle.kts the lines
     *      implementation("org.springframework.boot:spring-boot-starter-security")
     * 	    testImplementation("org.springframework.security:spring-security-test")
     * 	- from application.properties the lines
     * 	    spring.security.user.name=user
     *      spring.security.user.password=password
     *  - SpringSecurityConfig and SecurityController code and imports
     *  - from RoomControllerTest the lines
     *      import static org.mockito.BDDMockito.given;
     *      import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.*;
     *  - perhaps some other imports of the removed implementation of spring boot security that I've forgotten
     */

    // Spring object to mock call to our app
    @Autowired
    private MockMvc mockMvc;

    // The serializer used by Spring to send and receive data to/from the REST controller
    @Autowired
    private ObjectMapper objectMapper;

    // We choose to mock the DAO used in the REST controller to limit the scope of our test
    @MockBean
    private RoomDao roomDao;
    @MockBean
    private WindowDao windowDao;
    @MockBean
    private HeaterDao heaterDao;
    @MockBean
    private BuildingDao buildingDao;


    RoomEntity createRoomEntity(Long id, String name) {
        // Room is recreated before each test
        SensorEntity currentTemperature = new SensorEntity(SensorType.TEMPERATURE,"temperatureSensor");
        currentTemperature.setId(1L);
        currentTemperature.setValue(15.0);
        SensorEntity outsideTemperature = new SensorEntity(SensorType.TEMPERATURE, "Outside temperature");
        outsideTemperature.setId(2L);
        outsideTemperature.setValue(19.0);
        BuildingEntity building = new BuildingEntity("Building 1", outsideTemperature);
        RoomEntity roomEntity = new RoomEntity(1, name, currentTemperature,20.0, building);
        roomEntity.setId(id);
        return roomEntity;
    }

    @Test
    void shouldFindAll() throws Exception {
        Mockito.when(roomDao.findAll()).thenReturn(List.of(
                createRoomEntity(1L, "Room 1"),
                createRoomEntity(2L, "Room 2")
        ));

        mockMvc.perform(MockMvcRequestBuilders.get("/api/rooms").accept(APPLICATION_JSON))
                // check the HTTP response
                .andExpect(status().isOk())
                // the content can be tested with Json path
                .andExpect(
                        MockMvcResultMatchers
                                .jsonPath("[*].name")
                                .value(Matchers.containsInAnyOrder("Room 1", "Room 2"))
                );
    }

    @Test
    void shouldReturnNullWhenFindByUnknownId() throws Exception {
        Mockito.when(roomDao.findById(999L)).thenReturn(Optional.empty());

        mockMvc.perform(MockMvcRequestBuilders.get("/api/rooms/999").accept(APPLICATION_JSON))
                // check the HTTP response
                .andExpect(status().isOk())
                // the content can be tested with Json path
                .andExpect(MockMvcResultMatchers.content().string(""));
    }

    @Test
    void shouldFindById() throws Exception {
        RoomEntity roomEntity = createRoomEntity(1L, "Room 1");
        Mockito.when(roomDao.findById(999L)).thenReturn(Optional.of(roomEntity));

        mockMvc.perform(MockMvcRequestBuilders.get("/api/rooms/999").accept(APPLICATION_JSON))
                // check the HTTP response
                .andExpect(status().isOk())
                // the content can be tested with Json path
                .andExpect(MockMvcResultMatchers.jsonPath("$.name").value("Room 1"));
    }

    @Test
    void shouldNotUpdateUnknownEntity() throws Exception {
        RoomEntity roomEntity = createRoomEntity(1L, "Room 1");
        RoomCommand expectedRoom = new RoomCommand(roomEntity.getName(), roomEntity.getFloor(), roomEntity.getCurrentTemperature(), roomEntity.getTargetTemperature(), roomEntity.getWindows(), roomEntity.getHeaters(), roomEntity.getBuilding());
        String json = objectMapper.writeValueAsString(expectedRoom);

        Mockito.when(roomDao.findById(1L)).thenReturn(Optional.empty());

        mockMvc.perform(
                        MockMvcRequestBuilders
                                .put("/api/rooms/1")
                                .content(json)
                                .contentType(MediaType.APPLICATION_JSON_VALUE)
                )
                // check the HTTP response
                .andExpect(status().isBadRequest());
    }

    @Test
    void shouldUpdate() throws Exception {
        RoomEntity roomEntity = createRoomEntity(1L, "Room 1");
        RoomCommand expectedRoom = new RoomCommand(roomEntity.getName(), roomEntity.getFloor(), roomEntity.getCurrentTemperature(), roomEntity.getTargetTemperature(), roomEntity.getWindows(), roomEntity.getHeaters(), roomEntity.getBuilding());
        String json = objectMapper.writeValueAsString(expectedRoom);

        Mockito.when(roomDao.findById(1L)).thenReturn(Optional.of(roomEntity));

        mockMvc.perform(
                        MockMvcRequestBuilders
                                .put("/api/rooms/1")
                                .content(json)
                                .contentType(MediaType.APPLICATION_JSON_VALUE)
                )
                // check the HTTP response
                .andExpect(status().isOk())
                .andExpect(MockMvcResultMatchers.jsonPath("$.name").value("Room 1"))
                .andExpect(MockMvcResultMatchers.jsonPath("$.id").value("1"));
    }

    @Test
    void shouldCreate() throws Exception {
        RoomEntity roomEntity = createRoomEntity(1L, "Room 1");
        RoomCommand expectedRoom = new RoomCommand(roomEntity.getName(), roomEntity.getFloor(), roomEntity.getCurrentTemperature(), roomEntity.getTargetTemperature(), roomEntity.getWindows(), roomEntity.getHeaters(), roomEntity.getBuilding());
        String json = objectMapper.writeValueAsString(expectedRoom);

        Mockito.when(roomDao.existsById(1L)).thenReturn(false);
        Mockito.when(roomDao.save(Mockito.any(RoomEntity.class))).thenReturn(roomEntity);

        mockMvc.perform(
                        MockMvcRequestBuilders
                                .post("/api/rooms")
                                .content(json)
                                .contentType(MediaType.APPLICATION_JSON_VALUE)
                )
                // check the HTTP response
                .andExpect(status().isOk())
                .andExpect(MockMvcResultMatchers.jsonPath("$.name").value("Room 1"))
                .andExpect(MockMvcResultMatchers.jsonPath("$.id").value("1"));
    }

    @Test
    void shouldDelete() throws Exception {
        mockMvc.perform(MockMvcRequestBuilders.delete("/api/rooms/999"))
                .andExpect(status().isOk());
    }


    /*
    @Test
    @WithMockUser(username = "admin", password = "PASSWORD", roles = "ADMIN")
    void shouldLoadAWindowAndReturnNullIfNotFound() throws Exception {
        given(windowDao.findById(999L)).willReturn(Optional.empty());
        mockMvc.perform(get("/api/windows/999").accept(APPLICATION_JSON))
                // check the HTTP response
                .andExpect(status().isOk())
                // the content can be tested with Json path
                .andExpect(content().string(""));
    }
    */

}
